+++ audiomate/corpus/io/common_voice.py
@@ -103,7 +103,8 @@ class CommonVoiceReader(base.CorpusReader):
             issuer = assets.Speaker(idx, gender=gender, age_group=age)
             corpus.import_issuers(issuer)
             utterance = corpus.new_utterance(idx, idx, issuer.idx)
-            utterance.set_label_list(assets.LabelList.create_single(transcription, idx=assets.LL_WORD_TRANSCRIPT))
+            utterance.set_label_list(assets.LabelList.create_single(transcription,
+                                                                    idx=audiomate.corpus.LL_WORD_TRANSCRIPT))
 
             utt_ids.append(idx)