+++ tests/corpus/io/test_kaldi.py
@@ -3,6 +3,7 @@ import shutil
 import tempfile
 import unittest
 
+from audiomate import corpus
 from audiomate.corpus import io
 from audiomate.corpus import assets
 from tests import resources
@@ -94,17 +95,17 @@ class KaldiReaderTest(unittest.TestCase):
         utt_4 = ds.utterances['utt-4']
         utt_5 = ds.utterances['utt-5']
 
-        assert assets.LL_WORD_TRANSCRIPT in utt_1.label_lists.keys()
-        assert assets.LL_WORD_TRANSCRIPT in utt_2.label_lists.keys()
-        assert assets.LL_WORD_TRANSCRIPT in utt_3.label_lists.keys()
-        assert assets.LL_WORD_TRANSCRIPT in utt_4.label_lists.keys()
-        assert assets.LL_WORD_TRANSCRIPT in utt_5.label_lists.keys()
+        assert corpus.LL_WORD_TRANSCRIPT in utt_1.label_lists.keys()
+        assert corpus.LL_WORD_TRANSCRIPT in utt_2.label_lists.keys()
+        assert corpus.LL_WORD_TRANSCRIPT in utt_3.label_lists.keys()
+        assert corpus.LL_WORD_TRANSCRIPT in utt_4.label_lists.keys()
+        assert corpus.LL_WORD_TRANSCRIPT in utt_5.label_lists.keys()
 
-        assert len(utt_4.label_lists[assets.LL_WORD_TRANSCRIPT].labels) == 1
-        assert utt_4.label_lists[assets.LL_WORD_TRANSCRIPT].labels[0].value == 'who are they'
+        assert len(utt_4.label_lists[corpus.LL_WORD_TRANSCRIPT].labels) == 1
+        assert utt_4.label_lists[corpus.LL_WORD_TRANSCRIPT].labels[0].value == 'who are they'
 
-        assert utt_4.label_lists[assets.LL_WORD_TRANSCRIPT].labels[0].start == 0
-        assert utt_4.label_lists[assets.LL_WORD_TRANSCRIPT].labels[0].end == -1
+        assert utt_4.label_lists[corpus.LL_WORD_TRANSCRIPT].labels[0].start == 0
+        assert utt_4.label_lists[corpus.LL_WORD_TRANSCRIPT].labels[0].end == -1
 
 
 class KaldiWriterTest(unittest.TestCase):